refactor: MemberService를 Interface로 설계하여 이를 구현하도록 수정
